delete add widget modal
