+++ static/app/views/dashboardsV2/widgetBuilder/widgetBuilder.tsx
@@ -10,7 +10,6 @@ import {validateWidget} from 'sentry/actionCreators/dashboards';
 import {addErrorMessage, addSuccessMessage} from 'sentry/actionCreators/indicator';
 import {fetchOrgMembers} from 'sentry/actionCreators/members';
 import {loadOrganizationTags} from 'sentry/actionCreators/tags';
-import {generateOrderOptions} from 'sentry/components/dashboards/widgetQueriesForm';
 import DatePageFilter from 'sentry/components/datePageFilter';
 import EnvironmentPageFilter from 'sentry/components/environmentPageFilter';
 import * as Layout from 'sentry/components/layouts/thirds';
@@ -81,6 +80,7 @@ import {Header} from './header';
 import {
   DataSet,
   DEFAULT_RESULTS_LIMIT,
+  generateOrderOptions,
   getIsTimeseriesChart,
   getParsedDefaultWidgetQuery,
   getResultsLimit,
