+++ static/app/views/dashboardsV2/widgetCard/chart.tsx
@@ -262,7 +262,7 @@ class WidgetCardChart extends Component<WidgetCardChartProps, State> {
 
   chartComponent(chartProps): React.ReactNode {
     const {widget} = this.props;
-    const stacked = widget.queries[0].columns.length > 0;
+    const stacked = widget.queries[0]?.columns.length > 0;
 
     switch (widget.displayType) {
       case 'bar':
