+++ static/app/views/eventsV2/savedQuery/index.spec.tsx
@@ -1,15 +1,9 @@
 import {mountWithTheme} from 'sentry-test/enzyme';
-import {initializeOrg} from 'sentry-test/initializeOrg';
-import {render, screen, userEvent, waitFor} from 'sentry-test/reactTestingLibrary';
+import {render, screen, userEvent} from 'sentry-test/reactTestingLibrary';
 
-import {
-  openAddDashboardWidgetModal,
-  openAddToDashboardModal,
-} from 'sentry/actionCreators/modal';
 import {NewQuery} from 'sentry/types';
 import EventView from 'sentry/utils/discover/eventView';
 import {DisplayModes} from 'sentry/utils/discover/types';
-import {DashboardWidgetSource, DisplayType} from 'sentry/views/dashboardsV2/types';
 import DiscoverBanner from 'sentry/views/eventsV2/banner';
 import {ALL_VIEWS} from 'sentry/views/eventsV2/data';
 import SavedQueryButtonGroup from 'sentry/views/eventsV2/savedQuery';
@@ -309,119 +303,6 @@ describe('EventsV2 > SaveQueryButtonGroup', function () {
         expect.objectContaining({id: '1'})
       );
     });
-
-    it('opens a modal with the correct params for top 5 display mode', async function () {
-      const featuredOrganization = TestStubs.Organization({
-        features: ['dashboards-edit'],
-      });
-      const testData = initializeOrg({
-        ...initializeOrg(),
-        organization: featuredOrganization,
-      });
-      const savedTopNQuery = TestStubs.DiscoverSavedQuery({
-        display: DisplayModes.TOP5,
-        orderby: 'test',
-        fields: ['test', 'count()'],
-        topEvents: '2',
-      });
-      mount(
-        testData.router.location,
-        testData.organization,
-        testData.router,
-        EventView.fromSavedQuery(savedTopNQuery),
-        savedTopNQuery,
-        ['count()']
-      );
-      userEvent.click(screen.getByText('Add to Dashboard'));
-      expect(openAddDashboardWidgetModal).not.toHaveBeenCalled();
-
-      await waitFor(() => {
-        expect(openAddToDashboardModal).toHaveBeenCalledWith(
-          expect.objectContaining({
-            widget: {
-              title: 'Saved query #1',
-              displayType: DisplayType.AREA,
-              limit: 2,
-              queries: [
-                {
-                  aggregates: ['count()'],
-                  columns: ['test'],
-                  conditions: '',
-                  fields: ['test', 'count()', 'count()'],
-                  name: '',
-                  orderby: 'test',
-                },
-              ],
-            },
-            widgetAsQueryParams: expect.objectContaining({
-              defaultTableColumns: ['test', 'count()'],
-              defaultTitle: 'Saved query #1',
-              defaultWidgetQuery:
-                'name=&aggregates=count()&columns=test&fields=test%2Ccount()%2Ccount()&conditions=&orderby=test',
-              displayType: DisplayType.AREA,
-              source: DashboardWidgetSource.DISCOVERV2,
-            }),
-          })
-        );
-      });
-    });
-
-    it('opens a modal with the correct params for default display mode', async function () {
-      const featuredOrganization = TestStubs.Organization({
-        features: ['dashboards-edit'],
-      });
-      const testData = initializeOrg({
-        ...initializeOrg(),
-        organization: featuredOrganization,
-      });
-      const savedDefaultQuery = TestStubs.DiscoverSavedQuery({
-        display: DisplayModes.DEFAULT,
-        orderby: 'count()',
-        fields: ['test', 'count()'],
-        yAxis: ['count()'],
-      });
-      mount(
-        testData.router.location,
-        testData.organization,
-        testData.router,
-        EventView.fromSavedQuery(savedDefaultQuery),
-        savedDefaultQuery,
-        ['count()']
-      );
-      userEvent.click(screen.getByText('Add to Dashboard'));
-      expect(openAddDashboardWidgetModal).not.toHaveBeenCalled();
-
-      await waitFor(() => {
-        expect(openAddToDashboardModal).toHaveBeenCalledWith(
-          expect.objectContaining({
-            widget: {
-              title: 'Saved query #1',
-              displayType: DisplayType.LINE,
-              queries: [
-                {
-                  aggregates: ['count()'],
-                  columns: [],
-                  conditions: '',
-                  fields: ['count()'],
-                  name: '',
-                  // Orderby gets dropped because ordering only applies to
-                  // Top-N and tables
-                  orderby: '',
-                },
-              ],
-            },
-            widgetAsQueryParams: expect.objectContaining({
-              defaultTableColumns: ['test', 'count()'],
-              defaultTitle: 'Saved query #1',
-              defaultWidgetQuery:
-                'name=&aggregates=count()&columns=&fields=count()&conditions=&orderby=',
-              displayType: DisplayType.LINE,
-              source: DashboardWidgetSource.DISCOVERV2,
-            }),
-          })
-        );
-      });
-    });
   });
 
   describe('modifying a saved query', () => {
@@ -550,120 +431,6 @@ describe('EventsV2 > SaveQueryButtonGroup', function () {
       const buttonCreateAlert = wrapper.find(SELECTOR_BUTTON_CREATE_ALERT);
 
       expect(buttonCreateAlert.exists()).toBe(false);
-    });
-  });
-
-  describe('add dashboard widget', () => {
-    let initialData;
-    beforeEach(() => {
-      initialData = initializeOrg({
-        organization: {
-          features: ['discover-query', 'widget-viewer-modal', 'dashboards-edit'],
-          apdexThreshold: 400,
-        },
-        router: {
-          location: {query: {}},
-        },
-        project: 1,
-        projects: [],
-      });
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/events-stats/',
-        body: [],
-      });
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/',
-        body: [],
-      });
-    });
-
-    afterEach(() => {
-      MockApiClient.clearMockResponses();
-    });
-
-    it('opens widget modal when add to dashboard is clicked', () => {
-      mount(
-        initialData.router.location,
-        initialData.organization,
-        initialData.router,
-        errorsViewModified,
-        savedQuery,
-        ['count()']
-      );
-      userEvent.click(screen.getByText('Add to Dashboard'));
-      expect(openAddDashboardWidgetModal).toHaveBeenCalledWith(
-        expect.objectContaining({
-          defaultTableColumns: ['title', 'count()', 'count_unique(user)', 'project'],
-          defaultTitle: 'Errors by Title',
-          defaultWidgetQuery: {
-            conditions: 'event.type:error',
-            fields: ['count()'],
-            aggregates: ['count()'],
-            columns: [],
-            name: '',
-            orderby: '-count()',
-          },
-          displayType: 'line',
-        })
-      );
-    });
-
-    it('populates dashboard widget modal with saved query data if created from discover', () => {
-      mount(
-        initialData.router.location,
-        initialData.organization,
-        initialData.router,
-        errorsViewModified,
-        savedQuery,
-        yAxis
-      );
-      userEvent.click(screen.getByText('Add to Dashboard'));
-      expect(openAddDashboardWidgetModal).toHaveBeenCalledWith(
-        expect.objectContaining({
-          defaultTableColumns: ['title', 'count()', 'count_unique(user)', 'project'],
-          defaultTitle: 'Errors by Title',
-          defaultWidgetQuery: {
-            conditions: 'event.type:error',
-            fields: ['count()', 'failure_count()'],
-            aggregates: ['count()', 'failure_count()'],
-            columns: [],
-            name: '',
-            orderby: '-count()',
-          },
-          displayType: 'line',
-        })
-      );
-    });
-
-    it('adds equation to query fields if yAxis includes comprising functions', () => {
-      mount(
-        initialData.router.location,
-        initialData.organization,
-        initialData.router,
-        errorsViewModified,
-        savedQuery,
-        [...yAxis, 'equation|count() + failure_count()']
-      );
-      userEvent.click(screen.getByText('Add to Dashboard'));
-      expect(openAddDashboardWidgetModal).toHaveBeenCalledWith(
-        expect.objectContaining({
-          defaultTableColumns: ['title', 'count()', 'count_unique(user)', 'project'],
-          defaultTitle: 'Errors by Title',
-          defaultWidgetQuery: {
-            conditions: 'event.type:error',
-            fields: ['count()', 'failure_count()', 'equation|count() + failure_count()'],
-            aggregates: [
-              'count()',
-              'failure_count()',
-              'equation|count() + failure_count()',
-            ],
-            columns: [],
-            name: '',
-            orderby: '-count()',
-          },
-          displayType: 'line',
-        })
-      );
     });
   });
 });
